requirements.txt
manim
langchain
langchain-community
langgraph
langchain-mistralai
langchain-openai
pydantic
pypdf
pypdfium2
python-dotenv
faiss-cpu
